Fix socket read timeout

+++ config/settings.py
@@ -6,7 +6,7 @@ from json import load
 class Source:
     host: str
     port: str
-    timeout: int
+    timeout: float
 
 
 @dataclass
